+++ Frontend/src/app/pages/login/login/login.component.html
@@ -106,7 +106,7 @@
                 <p>
                     忘記密碼?
                 </p>
-                <a [routerLink]="['/' + path.forgotpassword]">找回密碼</a>
+                <a [routerLink]="['forgotpassword']">找回密碼</a>
             </div>
             <button (click)="login()">
                 登入
