+++ Frontend/src/app/pages/login/login/login.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 import { HttpClient } from '@angular/common/http';
-import { appPath } from '../../path/app-path-const';
+import { appPath } from '../../../path/app-path-const';
 
 @Component({
   selector: 'app-login',
