+++ Frontend/src/app/pages/register/verified/verified-routing.module.ts
