+++ Frontend/src/app/pages/register/verified/verified.component.spec.ts
