+++ Frontend/src/app/pages/register/verified/verified.module.ts
